requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
requests==2.31.0
mariadb==1.1.10
pillow==10.3.0
quart==0.19.4
werkzeug==3.0.1
aiofiles==23.2.1
aiohttp==3.9.1
psutil==5.9.8